Cleaning up

+++ RogueLike/Types.h
@@ -6,7 +6,7 @@
 using std::cout;
 using std::endl;
 
-const float TILE_SIZE = 64;
+const unsigned int TILE_SIZE = 64;
 const unsigned int WINDOW_WIDTH  = 860;
 const unsigned int WINDOW_HEIGHT = 640;
